Added HomeScreen and OTP Input Screen to Navigator
